Moved External Libraries directory, Refactoring SIMD

+++ src/bitmap/bitmap.filter.cpp
@@ -120,8 +120,8 @@ inline bool filter_bitmap (uint8_t* dest, const uint8_t* src, const dseed::size3
 				{
 					for (int fx = 0; fx < (int)mask.width; ++fx)
 					{
-						size_t cx = dseed::clamp<int> ((int)x + (fx - mask.width / 2), size.width - 1);
-						size_t cy = dseed::clamp<int> ((int)y + (fy - mask.height / 2), size.height - 1);
+						size_t cx = dseed::clamp<int> ((int)x + (fx - (int)mask.width / 2), size.width - 1);
+						size_t cy = dseed::clamp<int> ((int)y + (fy - (int)mask.height / 2), size.height - 1);
 
 						dseed::color_processor color = *((TPixel*)(src + depthZ + (stride * cy)) + cx);
 						color = color * mask.get_mask (fx, fy);
